backend/py-requirements.txt
numpy
requests
aiohttp
mysql-connector-python
